Put in new and shiny sprinker valves!  Using Toro TPV100, which can withstand 175 PSI, and burst at 1000PSI.
They also don't make a really really bad and loud noise!

Updated the button pushing to be a wail for thing rather then using the builtin debounce- as staic set this shit off.

Nothing like trying to launch something in barrell one, when it switches to 2 because of static.

Yeah!